+++ packages/database/prisma/migrations/20260708002923_m9_b3_product_sale_price/migration.sql
@@ -1,0 +1,4 @@
+-- AlterTable
+ALTER TABLE "Product" ADD COLUMN     "saleEndAt" TIMESTAMP(3),
+ADD COLUMN     "salePrice" INTEGER,
+ADD COLUMN     "saleStartAt" TIMESTAMP(3);